# Remove vinyl dependency