--- a/fraxl.cabal
+++ b/fraxl.cabal
@@ -84,16 +84,16 @@ test-suite examples
   ghc-options:         -Wall
   default-language:    Haskell2010
 
--- test-suite monadbench
---   type:                exitcode-stdio-1.0
---   hs-source-dirs:      tests
---   main-is:             MonadBench.hs
---   other-modules:       ExampleDataSource
---   build-depends:       base
---                      , fraxl
---                      , time
---   ghc-options:         -Wall -threaded -rtsopts -with-rtsopts=-N
---   default-language:    Haskell2010
+benchmark monadbench
+  type:                exitcode-stdio-1.0
+  hs-source-dirs:      tests
+  main-is:             MonadBench.hs
+  other-modules:       ExampleDataSource
+  build-depends:       base
+                     , fraxl
+                     , time
+  ghc-options:         -Wall -threaded -rtsopts -with-rtsopts=-N
+  default-language:    Haskell2010
 
 -- test-suite fraxl-test
 --   type:                exitcode-stdio-1.0