--- a/stack.yaml
+++ b/stack.yaml
@@ -4,5 +4,4 @@ packages:
 - '.'
 extra-deps:
 - text-1.2.2.1
-- vinyl-plus-0.1.1
-resolver: lts-6.1
+resolver: nightly-2017-12-01