This removes the dependency on vinyl which seems to have compilation errors with recent versions of GHC.