@ElvishJerricco 2018-02-13
Hey thanks! The reason I haven't done this myself is that we really need `Data.Union` to have the same performance as [`Data.OpenUnion`](https://hackage.haskell.org/package/freer-effects-0.3.0.1/docs/Data-OpenUnion.html) from `freer-effects`. Granted, the current implementation of `Data.Union` is not a regression from the `vinyl-plus` that `fraxl` used to use, but the performance of `freer-effects`'s union type was a pretty major innovation in the performance of extensible effects.

Also, I would prefer not to use `default-extensions` in the cabal file. It obscures information to the reader, and this is primarily a research package after all ;)