@ElvishJerricco 2018-02-13
Oh, I see you implemented `Data.Union` yourself here. I had assumed you'd used the `union` package, which I believe is identical to what you've made. Depending on that package would be better, especially since it has a module namespace conflict with this.