@gmalecha 2018-02-13
Thanks. I was having trouble with `vinyl` in my setup for some reason and `union` also uses it. The strictness annotations are a good point though. Even with that, `Data.Union` is probably a lot slower than the implementation from `freer-effects` (which is heavily optimized). Perhaps I will switch to that one.