@ElvishJerricco 2018-02-14
Yea, I think we need to have a conversation with the author of `union` to see if they'd be willing to sacrifice the nice, pure implementation of `union` in favor of the much faster implementation in `freer-effects`. If they're willing to do that, then we should just depend on `union`. But otherwise, we'll have to see if we can work with the `freer-effects` authors on getting their union factored out into a Hackage package somewhere.