@ElvishJerricco 2018-07-19
Closing this in favor of the `CoRec` type from `vinyl` until we find a constant time solution.