@MichaelXavier 2018-08-26
This stuff is all a bit above my pay grade but have you looked into [fastsum](https://github.com/patrickt/fastsum)? I think it does some TH to trade compile time for fast runtime. As long as you don't have to rebuild it, it is supposed to be pretty fast.