@ElvishJerricco 2018-08-27
Thad does seem quite good. I dislike taking the namespace `Data.Sum`, but that's minor. I'm not convinced that vinyl's CoRec couldn't be made to have faster operations that we need, but it  seems pretty nontrivial. So I would probably welcome a PR switching to `fastsum`.