@MichaelXavier 2018-08-27
I think its probably above my abilities to benchmark and add that. I'm sure what you have now is plenty fast for my needs, just wanted to see if you'd seen `fastsum` and if it would solve any problems.