@MichaelXavier 2018-08-28
I have a pretty tenuous understanding of the pieces involved but I was actually able to drop vinyl and replace it with fastsum and ran some benchmarks and saw a small performance increase in some cases.

On master:
```
stack bench --benchmark-arguments="seql 1000000"
# 18.82s before
stack bench --benchmark-arguments="seqr 1000000"
# 7.00s
stack bench --benchmark-arguments="tree 20"
# 19.17s
```

With fastsum:
```
stack bench --benchmark-arguments="seql 1000000"
# 15.79s
stack bench --benchmark-arguments="seqr 1000000"
# 7.08s
stack bench --benchmark-arguments="tree 20"
# 17.98s
```

This could be coincidental, the benchmarks are only running once, but I'll throw together a PR for you to review.