@christian-marie 2019-02-07
Can this please be released? Or the upper bound on vinyl relaxed?